fix offiste count bug;
